requirements.txt
tensorflow==2.16.2
pillow==11.2.1
fastapi==0.115.14
opencv-python==4.11.0.86
numpy==1.26.4
uvicorn==0.35.0
gunicorn==23.0.0
python-multipart
boto3
